% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_data_vs_fitted}
\alias{plot_data_vs_fitted}
\title{Visually Evaluate Model}
\usage{
plot_data_vs_fitted(model, ...)
}
\arguments{
\item{model}{A fitted model for which to plot its data vs its fitted values. Needs to be \code{\link[broom:reexports]{broom::augment()}}-able.}

\item{...}{Params passed to \code{\link[ggplot2:aes]{ggplot2::aes()}}. Should be the variables to plot, e.g. something like \verb{x = var1, y = var2}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} with two layers (points jittered in x-direction,
representing original values (in black) and fitted values (in red))
}
\description{
Plot values (in black) and their corresponding fitted values (in red),
according to \code{model}.
}
\examples{
plot_data_vs_fitted(lm(mpg ~ disp, data = mtcars), x = disp, y = mpg)
plot_data_vs_fitted(lm(mpg ~ disp + cyl + hp + wt, data = mtcars), x = wt, y = mpg)
}
