% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_track}
\alias{plot_track}
\title{Plot a Single Track with a Given Date}
\usage{
plot_track(track_date, points)
}
\arguments{
\item{track_date}{the date for which to plot the track}

\item{points}{A data.frame or tibble with (at least) a \code{date} and an sf geometry column containing a point for each observation}
}
\value{
a basic \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} with a \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} layer for the track
}
\description{
Takes a date, and a collection of points, then plots a track of all points which have the same
date as the one specified.
Points with \code{NA} values for latitude and/or longitude will be filtered out.
Remaining points will be converted \code{\link[=to_linestrings]{to_linestrings()}} before plotting.
Return value is the plot object, i.e. additional plot layers can be added in the usual ggplot way.
}
\examples{
plot_track(lubridate::ymd("2020-05-20"), track_details)
}
