% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_tracks}
\alias{summarize_tracks}
\title{Summarize a Collection of Detailed Track Points Grouped by Date}
\usage{
summarize_tracks(tracks)
}
\arguments{
\item{tracks}{A data.frame or tibble of track detail data points, e.g. \link{track_details}}
}
\value{
A tibble summarizing \code{tracks}, one row per date. Values (i.e. columns) are:
\itemize{
\item \code{date} the date of the trip
\item \code{weekday} weekday corresponding to \code{date}; factor with English weekday names; computed from \code{date}, only included for convenience
\item \code{distance_km} length of the track (in kilometers)
\item \code{time_min} duration of the track (in minutes)
\item \code{altitude_gain_m} total altitude gained during the track (in meters), without compensating for altitude lost
\item \code{temperature_c} median temperature of the track (in degrees Celsius)
\item \code{speed_km_h} average speed of the track (in km/h)
\item \code{avg_hr_bpm} average heart rate for the track (in beats per minute)
\item \code{below_zones_min} time spent below heart rate zone 1 (i.e. < 60\% of max. hr / < 109 bpm) (in minutes)
\item \code{zone1_min} time spent in heart rate zone 1 (i.e. [60;70)\% of max. hr / [109;127) bpm) (in minutes)
\item \code{zone2_min} time spent in heart rate zone 2 (i.e. [70;80)\% of max. hr / [127;145) bpm) (in minutes)
\item \code{zone3_min} time spent in heart rate zone 3 (i.e. [80;90)\% of max. hr / [145;164) bpm) (in minutes)
\item \code{zone4_min} time spent in heart rate zone 4 (i.e. [90;100)\% of max. hr / [164;182) bpm) (in minutes)
\item \code{above_zones_min} time spent above heart rate zone 4 (i.e. >= 100\% of max. hr / 182 bpm) (in minutes)
\item \code{samples} number of data point for the track, i.e. the number of data points which were aggregated for this row
}
}
\description{
Summarize a Collection of Detailed Track Points Grouped by Date
}
\examples{
# summarize all detailed data points by date
summarize_tracks(track_details)
# summarize only a specific date
library(dplyr)
track_details |>
  filter(date == "2020-06-01") |>
  summarize_tracks()
}
