% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_cantons}
\alias{swiss_cantons}
\title{Swiss cantons}
\format{
\subsection{\code{swiss_cantons}}{

A list with 26 rows and 15 columns:
\describe{
\item{KTNR}{canton number}
\item{KTNAME}{canton name}
\item{GRNR}{number of the greater region the canton belongs to}
\item{AREA_HA}{area in hectares (rounded to whole numbers)}
\item{E_MIN}{canton's minimum eastern coordinate value}
\item{E_MAX}{canton's maximum eastern coordinate value}
\item{N_MIN}{canton's minimum northern coordinate value}
\item{N_MAX}{canton's maximum northern coordinate value}
\item{E_CNTR}{canton's centroid coordinate (east)}
\item{N_CNTR}{canton's centroid coordinate (north)}
\item{Z_MIN}{canton's minimum height coordinate value}
\item{Z_MAX}{canton's maximum height coordinate value}
\item{Z_AVG}{canton's average height coordinate value}
\item{Z_MED}{canton's median height coordinate value}
\item{geometry}{\link[sf:sfc]{sf::sfc_MULTIPOLYGON} tracing the canton's borders}
}
}
}
\source{
\url{https://www.bfs.admin.ch/bfs/de/home/dienstleistungen/geostat/geodaten-bundesstatistik/administrative-grenzen/generalisierte-gemeindegrenzen.assetdetail.30487000.html}
}
\usage{
swiss_cantons
}
\description{
2024 geographical borders of all 26 Swiss cantons as provided by the Swiss Federal Statistical Office (Bundesamt für Statistik, BfS).
This is a simple feature (sf) object with additional information.
It can be used directly with \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} to draw a map of the borders of the cantons.
The coordinate reference system (CRS) is the Swiss LV95 system (EPSG 2056).
}
\keyword{datasets}
