% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_lakes}
\alias{swiss_lakes}
\title{Swiss lakes}
\format{
\subsection{\code{swiss_lakes}}{

A list with 23 rows and 10 columns:
\describe{
\item{GMDNR}{lake number}
\item{GMDNAME}{lake name}
\item{SEE_HA}{area in hectares (rounded to whole numbers)}
\item{E_MIN}{lake's minimum eastern coordinate value}
\item{E_MAX}{lake's maximum eastern coordinate value}
\item{N_MIN}{lake's minimum northern coordinate value}
\item{N_MAX}{lake's maximum northern coordinate value}
\item{E_CNTR}{lake's centroid coordinate (east)}
\item{N_CNTR}{lake's centroid coordinate (north)}
\item{geometry}{\link[sf:sfc]{sf::sfc_MULTIPOLYGON} tracing the lakeshore}
}
}
}
\source{
\url{https://www.bfs.admin.ch/bfs/de/home/dienstleistungen/geostat/geodaten-bundesstatistik/administrative-grenzen/generalisierte-gemeindegrenzen.assetdetail.30487000.html}
}
\usage{
swiss_lakes
}
\description{
2024 map geometry of the 23 largest Swiss lakes as provided by the Swiss Federal Statistical Office (Bundesamt für Statistik, BfS).
This is a simple feature (sf) object with additional information.
It can be used directly with \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} to draw a map of the Swiss lakes.
The coordinate reference system is the Swiss LV95 system (EPSG 2056).
}
\keyword{datasets}
