% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{track_classes}
\alias{track_classes}
\title{Categorized road bike trips from 2018 to 2023}
\format{
\subsection{\code{track_classes}}{

A data frame with 157 rows and 3 columns:
\describe{
\item{date}{Date of the trip (date)}
\item{track}{Classification of the track (by route taken) (integer)}
\item{direction}{Direction in which the route was driven; only available for \code{track} == 3 (factor)}
}
}
}
\usage{
track_classes
}
\description{
Tracks categorized by route and direction. Routes that were only driven in one
direction have no direction information (\code{direction} == NA),
routes that were only driven once might not have their own class (\code{track} == NA).
}
\details{
Tracks are identified by \code{date}. Only contains track class (\code{track}) and
direction (\code{direction}), for more information regarding tracks use
\code{tracks} data, or \code{track_details}.
}
\keyword{datasets}
