% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracks}
\alias{tracks}
\title{Summary road bike trip data from 2018 to 2023}
\format{
\subsection{\code{tracks}}{

A data frame with 157 rows and 15 columns:
\describe{
\item{date}{Date of the trip (date)}
\item{weekday}{Weekday of the trip; can be calculated from \code{date} using \code{\link[base:weekday.POSIXt]{base::weekdays()}}, but is included directly for convenience (factor)}
\item{distance_km}{Length of the trip in kilometers (integer)}
\item{time_min}{Trip duration in minutes (integer)}
\item{altitude_gain_m}{Cumulated total altitude gain during the trip, only counting positive gains (all trips have same start and finish points, i.e. overall gain is 0 when including negative gains)  (integer)}
\item{temperature_c}{Median temperature in °C (integer)}
\item{speed_km_h}{Average speed in km/h (integer)}
\item{avg_hr_bpm}{Average heart rate in beats per minute (integer)}
\item{below_zones_min}{Time below heart rate zone 1 in minutes, i.e. heart rate < 109 bpm (integer)}
\item{zone1_min}{Time in heart rate zone 1 in minutes, i.e. heart rate within [60;70)\% of configured maximum (182 bpm) = [109;127) bpm (integer)}
\item{zone2_min}{Time in heart rate zone 2 in minutes, i.e. heart rate within [70;80)\% of configured maximum (182 bpm) = [127;145) bpm (integer)}
\item{zone3_min}{Time in heart rate zone 3 in minutes, i.e. heart rate within [80;90)\% of configured maximum (182 bpm) = [145;164) bpm (integer)}
\item{zone4_min}{Time in heart rate zone 4 in minutes, i.e. heart rate within [90;100)\% of configured maximum (182 bpm) = [164;182) bpm (integer)}
\item{above_zones_min}{Time above heart rate zone 4 in minutes, i.e. heart rate >= 182 bpm (integer)}
\item{samples}{Number of datapoints, i.e. for a given trip, \code{track_details} countains \code{samples} rows for that trip (integer)}
}
}
}
\usage{
tracks
}
\description{
Aggregated summary track data for road bike trips from 2018 to 2023, one row per trip.
Details for each trip are in \code{track_details}.
}
\keyword{datasets}
